% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hvtiRdatabuild-package.R
\docType{package}
\name{hvtiRdatabuild-package}
\alias{hvtiRdatabuild}
\alias{hvtiRdatabuild-package}
\title{hvtiRdatabuild: Build and Verify Analytic Datasets for the HVTI CORR Group}
\description{
Builds analysis-ready clinical datasets for the clinical investigations statistics group within The Heart and Vascular Institute at the Cleveland Clinic, and verifies them against the legacy 'SAS' datasets they replace.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ehrlinger/hvtiRdatabuild}
  \item \url{https://ehrlinger.github.io/hvtiRdatabuild/}
  \item Report bugs at \url{https://github.com/ehrlinger/hvtiRdatabuild/issues}
}

}
\author{
\strong{Maintainer}: John Ehrlinger \email{john.ehrlinger@gmail.com}

Authors:
\itemize{
  \item John Ehrlinger \email{john.ehrlinger@gmail.com}
}

}
\keyword{internal}
