% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dw_pull.R
\name{print.pull_result}
\alias{print.pull_result}
\title{Print a pull result}
\usage{
\method{print}{pull_result}(x, ...)
}
\arguments{
\item{x}{A \code{pull_result} from \code{\link[=dw_pull]{dw_pull()}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a pull result
}
